rename directory name safety_check to path_safety_checker
