+++ planning/behavior_path_planner/test/test_safety_check.cpp
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 #include "behavior_path_planner/marker_utils/utils.hpp"
-#include "behavior_path_planner/utils/safety_check/safety_check.hpp"
+#include "behavior_path_planner/utils/path_safety_checker/safety_check.hpp"
 #include "tier4_autoware_utils/tier4_autoware_utils.hpp"
 
 #include <geometry_msgs/msg/pose.hpp>
